To display all the lines where the word under the cursor occurs, simply do in Normal mode
[I. This can be useful to find a count of lines of search occurrences. Each line displayed is numbered.
In order to jump to the <n>th line of occurrence, do: <n>[I; this is similar to <n>[i, which displays the <n>th line of occurrence.
The function and mappings below work in visual mode too, so that the search will be done for the visual highlight.
In addition, [I without <n> asks for the occurrence number to jump to.